# docs/OpenAPI-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "F24 | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "F24"
        ],
        "responses": { "302": { "description": "Redirect to root folder" } }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "F24"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/health/db": {
      "get": {
        "tags": [
          "F24"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "503": {
            "description": "Database unavailable"
          }
        }
      }
    },
    "/files/{id}": {
      "delete": {
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "404": { "description": "Not Found" }
        }
      }
    },
    "/folders/{id}": {
      "get": {
        "tags": [
          "Folders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FolderContentsDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderContentsDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderContentsDto"
                }
              }
            }
          },
          "404": { "description": "Not Found" }
        }
      },
      "post": {
        "tags": [
          "Folders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEntryRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEntryRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EntryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryDto"
                }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "409": { "description": "Conflict" }
        }
      },
      "delete": {
        "tags": [
          "Folders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "404": { "description": "Not Found" },
          "409": { "description": "Conflict" }
        }
      }
    },
    "/search": {
      "get": {
        "tags": [
          "Search"
        ],
        "parameters": [
          {
            "name": "prefix",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "folder",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SearchMode",
              "default": "PrefixAll"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/SearchResultDto" }
                }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateEntryRequest": {
        "required": [
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/EntryType"
          }
        }
      },
      "EntryDto": {
        "required": [
          "id",
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "EntryType": {
        "type": "string",
        "enum": ["Folder", "File"]
      },
      "FolderContentsDto": {
        "required": [
          "id",
          "name",
          "parentId",
          "path",
          "children"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "parentId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "path": {
            "type": "string"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntryDto"
            }
          }
        }
      },
      "SearchMode": {
        "type": "string",
        "enum": ["PrefixAll", "ExactCurrent"]
      },
      "SearchResultDto": {
        "required": ["id", "name", "path", "parentId"],
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "path": { "type": "string" },
          "parentId": { "type": "string", "format": "uuid" }
        }
      }
    }
  },
  "tags": [
    {
      "name": "F24"
    },
    {
      "name": "Files"
    },
    {
      "name": "Folders"
    },
    {
      "name": "Search"
    }
  ]
}